Initial commit of new proposal page
