+++ apps/frontend/src/shared-research/ResearchOutput.tsx
@@ -1,8 +1,4 @@
-import {
-  NotFoundPage,
-  SharedResearchProposal,
-  SharedResearchOutput,
-} from '@asap-hub/react-components';
+import { NotFoundPage, SharedResearchOutput } from '@asap-hub/react-components';
 import { sharedResearch, useRouteParams } from '@asap-hub/routing';
 
 import { useBackHref } from '../hooks';
@@ -19,11 +15,7 @@ const ResearchOutput: React.FC = () => {
   if (researchOutputData) {
     return (
       <Frame title={researchOutputData.title}>
-        {researchOutputData.type === 'Proposal' ? (
-          <SharedResearchProposal {...researchOutputData} backHref={backHref} />
-        ) : (
-          <SharedResearchOutput {...researchOutputData} backHref={backHref} />
-        )}
+        <SharedResearchOutput {...researchOutputData} backHref={backHref} />
       </Frame>
     );
   }
