+++ packages/model/src/research-output.ts
@@ -11,6 +11,7 @@ export const researchOutputTypes = [
   'Protocol',
   'Lab Resource',
   'Article',
+  'Grant Document',
 ] as const;
 
 export type ResearchOutputType = typeof researchOutputTypes[number];
@@ -94,6 +95,7 @@ export type ResearchOutputResponse = {
 
 export const researchOutputLabels: Record<ResearchOutputType, string> = {
   Proposal: 'Open External Link',
+  'Grant Document': 'Open External Link',
   Presentation: 'View on Google',
   Dataset: 'Open External Link',
   Bioinformatics: 'Open External Link',
