+++ packages/react-components/src/index.ts
@@ -147,7 +147,6 @@ export {
   SharedResearchList,
   SharedResearchOutput,
   SharedResearchPage,
-  SharedResearchProposal,
   SigninForm,
   SigninPage,
   ExpertiseAndResourcesModal,
